djangoApi/requirements.txt
Django==4.2.11
djangorestframework
psycopg2
django-cors-headers
